Remove the notion of "lite mode" completely
As this disables everything from sporks to masternodes it has not been
completely compatible with the network to run in this mode for quite some
time. To avoid having to rewrite parts and add exceptions all over the
code base for this mode it's better to just remove support for it.