Sometimes the StakeMiner thread could stall the daemon on shutdown
During execution, the stake miner checks for mintable inputs. If none
are available it stops, waiting for mintable inputs to be available.
During shutdown of the daemon, mintable inputs are cleared, sometimes
causing the staking thread to get stuck inside this loop.

Fixed by adding a fShutdown check inside the loop that checks for mintable
inputs.