Implement a misbehavior log under each node/peer
For now, this log can be viewed via the "getpeerinfo" RPC command. It will
later be viewable under the "listbanned" command as well, but we need to
implement it and add serialization and deserialization to the banlist file
that the daemon saves.

These changes also clean up many debug messages and reformat some of the
code.