Implement initial version of block index class
This class will implement a moving cache of block indexes - only holding
the 1024 most recently used blocks in memory - reverting to the
underlying LevelDB database whenever a block does not exist in memory.

This is just an initial implementation - it does not work and does not
compile properly. It has only been committed for reference thus far.