Rework the BlockIndex class to work with CDiskBlockIndex instances
This suits us a lot better as we don't really need a double linked list
when our in-memory data structure is so small.

This change actually simplifies the code more than anything. Eventually
we will be able to throw away the base CBlockIndex class and just work
with the "disk version" of tje class.